+++ src/actions/ens/getEnsAddress.bench.ts
@@ -1,12 +1,17 @@
-import { bench, describe } from 'vitest'
+import { beforeAll, bench, describe } from 'vitest'
 
 import {
   ethersProvider,
   ethersV6Provider,
   publicClient,
+  setBlockNumber,
 } from '../../_test/index.js'
 
 import { getEnsAddress } from './getEnsAddress.js'
+
+beforeAll(async () => {
+  await setBlockNumber(16773780n)
+})
 
 describe('Get ENS Name', () => {
   bench('viem: `getEnsAddress`', async () => {