+++ src/actions/ens/getEnsName.bench.ts
@@ -1,12 +1,16 @@
-import { bench, describe } from 'vitest'
+import { beforeAll, bench, describe } from 'vitest'
 
 import {
   ethersProvider,
   ethersV6Provider,
   publicClient,
+  setBlockNumber,
 } from '../../_test/index.js'
-
 import { getEnsName } from './getEnsName.js'
+
+beforeAll(async () => {
+  await setBlockNumber(16773780n)
+})
 
 describe('Get ENS Name', () => {
   bench('viem: `getEnsName`', async () => {